Allow non-composite title
The title should not require a two-part name consisting of both parts of
the namevar. The title should be able to be a simple resource title
provided that the user supplies name and target in the parameters of the
resource declaration.

This commit modifies the title_patterns method to allow simple name
titles in addition to composite name:target titles.