Clean up PKCS12 generation and testing
The `to_pkcs12` method was declaring a `Tempfile` object which was being
closed by the calling object. This is messy. The declaration of the
temp file is moved to the `import_ks` method where it was originally
without losing the fixes of #10.

The specs around these to methods are also enhanced to declare their
respective duties in a less tightly-coupled fashon.