Don't assume FOSS paths in java_ks tests
Previously we were using /etc/puppet as the root of the ssl dir and the
module path. This moves the tests to use the default nodes 'puppetpath'
and 'distmoduledir' respectively.  Tests were these frequently used take
advantage of RSpec's `let' to set them to their Puppet config name
aliases "confdir" and "modulepath" respectively.