Stop depending on puppet certs to test java_ks
Previously we removed the call to `puppet master` during acceptance
testing setup because it caused conflicts in agent-master multinode
setups. However, we realized that the reason for the call to `puppet
master` in the first place was so that java_ks could use the existing
puppet certs that exist on the puppet master. This patch refactors the
tests to generate a CA and private key itself. This avoids having tests
depend on a certain node configuration.

It also adds /usr/bin to the PATH on solaris so that the java_ks type
can use the openssl command in tests on solaris.