Fix solaris 10 tests
As long as basic_spec.rb runs before the other acceptance tests, we
don't need to include the java class in the other tests. This patch
removes the java class from the other tests since it is not supported
on solaris 10. It also fixes the PATH so that we can find ruby on
solaris 10, foss or PE.