(MODULES-4605) (PE-17635) Wrap keytool in timeout
Thanks to @MercateoITO in GH PR #176

The keytool may hang on occasion and block on IO in an unreproducable
case of the tempfile being removed before the keytool exits. We can wrap
the keytool execution in a timeout to handle this case such that it will
not eternally hang and block further puppet runs.