(MODULES-5117) Fix java install on windows in acceptance
This commit fixes installing java on windows in the acceptance tests
by moving the installation logic to the spec_helper_acceptance so that
java is always installed when needed, and removes basic_spec.rb because
having the installation logic in one spec file means all spec files must
be run in order to have a properly setup environment.

This commit also migrates the java install from using windows_java to
puppetlabs-chocolatey. This follows a standard at Puppet to use chocolatey
in acceptance to setup an environment.