(MODULES-5357) Pin JDK installation pacakge to 8.0.144
Previously the acceptance tests installed JDK from the chocolatey pacakge repo
however the version of the installed pacakges was not enforced and the
calculation for the Java path then fails as it is expected java 8u131.  This
commit pins the JDK8 pacakged to the latest 8u141 in the chocolatey repo and
updates the java version logic with the new version number.