(FM-6547) Fix shell string in pkcs12 test
The shell string in the it block of the pkcs12 acceptance tests does
not include escaped quoting around the path to `keytool` prior to
this commit. This causes a failure on windows because the path includes
a space. This commit wraps the path to keytool in escaped quotes to
resolve the issue.