(FM-6547) Fix unit tests on windows
Prior to this commit, the unit tests for java_ks could never pass
on windows due to use of linux paths and commands hard-coded. This
commit adds a variable for the temp paths in both spec test files
and guards against running tests that rely on linux commands on a
windows machine.