(maint) Ignore JDK keytool warnings in tests for DER certs
On Windows, the keystore command warns about using a proprietary format when using DER formatted certs. We should not take this as a failure, but also, we should also not blindly ignore all STDERR from the result either. If we get an exit code of 1, we'll check to see if the STDERR message was the cert format warning and still pass the test. We will still catch any errors that occur and are not related. This issue was introduced after https://github.com/puppetlabs/puppetlabs-java_ks/pull/319 was merged.

Fix Rubocop violations