Change latest/current comparison to account for chains
When certificate chains are being compared, the 'latest' one can have
several SHA1 checksums in them, which the code will flatten and join
into a single string. This gets compared against what is in the
keystore, which only has a single SHA1 checksum in it. This results in a
failed check every time.

Instead of checking for equality between those two values, we can just
check to see if 'latest' contains 'current'.