Add support for certificate_content and private_key_content parameters
The current implementation only allows you to pass a file name to the certificate
and private_key parameters. When you are fetching certificates from vault or another secure store,
you'll first have to save them to a file. This is very innconveniant.

This PR add's the parameters certificate_content and private_key_content. These parameters
are mutually exclusive from their file counterparts.

With this change, you can now fetch a certificate and/or a password from vault
(through a hiera lookup for example) and use it directly on the type.

Because these values can be sensitive, both of the new parameters support
passing the value as a sensitive data type.